Fix ID overflow issue and code style (use Long instead of Int)

+++ app/src/main/java/com/example/todolist/Todo.kt
@@ -9,8 +9,7 @@ import java.util.Date
  * Contains only data, no business logic.
  */
 data class Todo(
-    var id: Int,
+    var id: Long,
     var title: String,
     var createdAt: Date
-
 )
